Debug because of merge

+++ app/Http/Controllers/ScenarioController.php
@@ -11,6 +11,7 @@ use App\Models\ScenarioStep;
 use App\Models\Mockup;
 use App\Models\Event;
 use App\Models\AcknowledgedEvent;
+use App\Models\Project;
 use Illuminate\Support\Facades\Auth;
 
 class ScenarioController extends Controller
@@ -19,7 +20,8 @@ class ScenarioController extends Controller
   function show($projectId, $scenarioId)
   {
     $scenario = Scenario::find($scenarioId);
-    return view('scenario.show', ['projectId'=>$projectId, 'scenario'=>$scenario]);
+    $project = Project::find($projectId);
+    return view('scenario.show', ['projectId'=>$projectId, 'scenario'=>$scenario, 'mockups' => $project->mockups]);
   }
 
   //update scenario
@@ -124,13 +126,13 @@ class ScenarioController extends Controller
       if($request->file('maquette')->isValid()){
         $file = $request->file('maquette');
         $newName = uniqid('img').".".$file->getClientOriginalExtension();
-        $path = $file->move("mockups/$projectid/$scenarioId", $newName);
+        $path = $file->move("mockups/$projectid/", $newName);
 
-        $scenario = Scenario::find($scenarioId);
+        $project = Project::find($projectid);
 
         $mockup = new Mockup;
         $mockup->url = $newName;
-        $mockup->scenario()->associate($scenario);
+        $mockup->project()->associate($project);
         $mockup->save();
 
       }
@@ -154,8 +156,10 @@ class ScenarioController extends Controller
 
     if(!is_null($image)){
       $filename = "mockups/$projectid/".$image->url;
-      File::delete($filename);
-      $image->delete();
+
+      if($image->delete())
+        File::delete($filename);
+
     }
   }
 }
